refactored to make items for the WarehouseList.js
