+++ client/src/components/header/header.js
@@ -20,12 +20,12 @@ class Header extends Component {
           <ul className="header__list">
             <li className="header__item header__item--margin">
               <Link
-                to="/warehouse"
+                to="/warehouses"
                 className={`header__list-button ${
-                  pathname === "/warehouse" ? "header__list-button--active" : ""
+                  pathname === "/warehouses" ? "header__list-button--active" : ""
                 }`}
               >
-                Warehouse
+                Warehouses
               </Link>
             </li>
             <li className="header__item">
